LayoutTests/platform/glib/fast/text/international/system-language/navigator-language/navigator-language-en-expected.txt

FAIL navigator.language should be en-US. Was en.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
